Move IOC logics to extension method
